requirements.txt
sqlalchemy
databases
pydantic
fastapi
passlib
python-jose
uvicorn
asyncpg
psycopg2-binary
websockets
ocpp
fastapi[all]